mgr/volumes: Fix subvolume removal
When subvolumegroup quota is set, subvolume removal
fails with EXDEV error. The subvolume deletion, renames
the subvolume directory to the trash directory which
resides outside the subvolume group directory resulting
in EXDEV error.

The patches solves the issue by changing the subvolume
trash directory location to '/volumes/<group>/.trash' when
the group quota is set.

Fixes: https://tracker.ceph.com/issues/53509
Signed-off-by: Kotresh HR <khiremat@redhat.com>