qa: Fix cephfs_client_version
The multifs auth caps bug has a fix both in client and mds.
If it's old client and not patched, we expect that the fs
with 'rw' would end up having 'r' caps with the multifs
auth caps used as in the test
'test_multifs_single_client_cross_access_r_caps_end'.
This patch adds the conditional to validate the same.

This commit makes use of the PR #64005

Signed-off-by: Kotresh HR <khiremat@redhat.com>